+++ db/src/repository/SQSDatabase.cpp
@@ -1558,6 +1558,7 @@ namespace AwsMock::Database {
                     }
                 }
                 session.commit_transaction();
+                return;
 
             } catch (const mongocxx::exception &exc) {
                 session.abort_transaction();
@@ -1565,6 +1566,7 @@ namespace AwsMock::Database {
                 throw Core::DatabaseException(exc.what());
             }
         }
+        _memoryDb.AdjustMessageCounters();
     }
 
 }// namespace AwsMock::Database